requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
faiss-cpu
sentence-transformers
python-dotenv
